Move base test classes to own module.
This prevents the installation of unwanted dependencies in project which need to import the news builders.
