+++ docs/HISTORY.txt
@@ -5,7 +5,7 @@ Changelog
 1.8.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Move base test classes to own module. [mbaechtold]
 
 
 1.8.3 (2017-03-30)
