+++ ftw/news/testing.py
@@ -9,6 +9,8 @@ from plone.app.testing import FunctionalTesting
 from plone.app.testing import PloneSandboxLayer
 from plone.testing import z2
 from zope.configuration import xmlconfig
+from ftw.simplelayout.tests import builders
+from ftw.subsite.tests import builders
 import ftw.news.tests.builders
