+++ ftw/news/tests/test_news_archive_portlets.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from ftw.builder import Builder
 from ftw.builder import create
 from ftw.news.testing import FTW_NEWS_FUNCTIONAL_TESTING
-from ftw.news.tests import FunctionalTestCase
+from ftw.news.tests.base import FunctionalTestCase
 from ftw.testbrowser import browsing
 from plone import api
 import transaction
